Initial barebones UI development for Skeletal Increment
A barebones implementation of the UI needed for the skeletal increment. Prepares JSON packages for move, accuse, and suggest actions. Also interprets the packages and prints a human-readable version of them.

Still TODO: Repackage payloads according to the standardized game payload, update interpretation on that end as well, connect to server.