Full implementation of messaging
Resolved UI issues. This uses a different message management system than the messagequeue system that was previously implemented. We'll have to discuss the purpose of messagequeue and how to either merge it into the new system, or keep it as is.

Also, the messaging system made it incredibly easy to implement a basic chat functionality between players, so that is implemented on this branch as well. Can easily be removed if we don't want to add this.